RTF Macro Sorter v.2.5.1
This version only works with the new version of the RTF Report, titled
"Daily RTF Report".

We are phasing out support for the older versions of the RTF report,
which are titled "[COD] RTF daily report1" (where [COD] would be your
branch code), and therefore will not be assigning resources to make the
macro file work with that version.